refactor(reddit-conversions-api): drop the v2 file split, restore utils.ts/types.ts
Deletes types-v2.ts and utils-v2.ts, and restores utils.ts/types.ts to
their original main-branch content verbatim. v2 stays completely
untouched; v3 is purely additive in types-v3.ts/utils-v3.ts, which now
also owns the FLAGON_NAME/isCanary/resolveVersion routing helpers.

standardEvent/customEvent call the original send() for the v2 path and
sendV3() for v3, per resolveVersion(payload.api_version, features).

Known follow-up: utils.ts still references settings.test_mode, which
no longer exists on Settings (removed earlier). Left as-is per
instruction not to worry about lint/type checks for this push.
